pyproject.toml
[project]
name = "zhou-v2"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10"
dependencies = [
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
